% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{dec_to_hex}
\alias{dec_to_hex}
\title{Convert decimal to hexadecimal.}
\usage{
dec_to_hex(dec)
}
\arguments{
\item{dec}{Decimal number.}
}
\value{
String.
}
\description{
Convert decimal to hexadecimal.
}
\examples{
dec_to_hex(21106)
}
