% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpc-eth.R
\name{eth_accounts}
\alias{eth_accounts}
\title{Returns addresses owned by client.}
\usage{
eth_accounts()
}
\value{
A vector of 20 byte (hexadecimal) Ethereum addresses.
}
\description{
Returns addresses owned by client.
}
\examples{
\dontrun{
eth_accounts()
}
}
