% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpc-eth.R
\name{eth_blockNumber}
\alias{eth_blockNumber}
\title{Returns the number of most recent block.}
\usage{
eth_blockNumber()
}
\value{
Integer.
}
\description{
Returns the number of most recent block.
}
\examples{
\dontrun{
eth_blockNumber()
}
}
