% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpc-eth.R
\name{eth_call}
\alias{eth_call}
\title{Executes a new message call immediately without creating a transaction on the block chain.}
\usage{
eth_call(
  from_address = NULL,
  to_address,
  gas = NULL,
  gasPrice = NULL,
  value = NULL,
  data = NULL,
  number = "latest"
)
}
\description{
Not completed just to_address and data
https://eth.wiki/json-rpc/API#eth_accounts
}
\examples{
\dontrun{
#eth_call(to_address =token_address,data = data)
#eth_call(to_address ="0x0D8775F648430679A709E98d2b0Cb6250d2887EF",
#data = "0x70a08231000000000000000000000000D79D326318aa8aBd5B84984dd5F258f7B5ecD356")
}
}
