% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth.R
\name{eth_coinbase}
\alias{eth_coinbase}
\title{Returns the client coinbase address.}
\usage{
eth_coinbase()
}
\value{
A 20 byte (hexadecimal) Ethereum address.
}
\description{
Returns the client coinbase address.
}
\examples{
eth_coinbase()
}
