% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth.R
\name{eth_getBalance}
\alias{eth_getBalance}
\title{Returns the balance of the account at specified address.}
\usage{
eth_getBalance(address = NULL, block = "latest")
}
\arguments{
\item{address}{A 20 byte (hexadecimal) Ethereum address.}

\item{block}{An integer block number (as a hexidecimal string) or one of "earliest", "latest" or "pending".}
}
\description{
Returns the balance of the account at specified address.
}
