% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth.R
\name{eth_getBlock}
\alias{eth_getBlock}
\title{Returns information about a specified block.}
\usage{
eth_getBlock(hash = NULL, number = "latest", full = TRUE)
}
\arguments{
\item{hash}{Hash of a block.}

\item{number}{An integer block number (as a hexidecimal string) or one of "earliest", "latest" or "pending".}

\item{full}{Whether or not to return full transaction objects.}
}
\value{
List.
}
\description{
Combines functionality from eth_getBlockByHash() and eth_getBlockByNumber().
}
\examples{
eth_getBlock()
eth_getBlock("0xb6d656ead4c3d4b1aa24d6b4d3d4cde8c090794e597258993512d650f088fcba")
eth_getBlock("0x4720FF")
}
