% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth.R
\name{eth_getBlockByHash}
\alias{eth_getBlockByHash}
\title{Returns information about a specified block.}
\usage{
eth_getBlockByHash(hash, full = TRUE)
}
\arguments{
\item{hash}{Hash of a block.}

\item{full}{Whether or not to return full transaction objects.}
}
\value{
List.
}
\description{
Returns information about a specified block.
}
\examples{
eth_getBlockByHash("0x8575df1eb3df61f3880628ca8e495038ee0b278c0aa48fe41f80b8d0d4e83e79")
}
