% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth.R
\name{eth_getBlockByNumber}
\alias{eth_getBlockByNumber}
\title{Returns information about a specified block.}
\usage{
eth_getBlockByNumber(block = "latest", full = TRUE)
}
\arguments{
\item{block}{An integer block number (as a hexidecimal string) or one of "earliest", "latest" or "pending".}

\item{full}{Whether or not to return full transaction objects.}
}
\value{
List.
}
\description{
Returns information about a specified block.
}
\examples{
eth_getBlockByNumber()
eth_getBlockByNumber("0x4720FE")
}
