% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth.R
\name{eth_getBlockTransactionCountByHash}
\alias{eth_getBlockTransactionCountByHash}
\title{Returns the number of transactions in a block from a block matching the given block hash.}
\usage{
eth_getBlockTransactionCountByHash(hash)
}
\arguments{
\item{hash}{Hash of a block.}
}
\description{
Returns the number of transactions in a block from a block matching the given block hash.
}
