% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth.R
\name{eth_getBlockTransactionCountByNumber}
\alias{eth_getBlockTransactionCountByNumber}
\title{Returns the number of transactions in a specified block.}
\usage{
eth_getBlockTransactionCountByNumber(block = "latest")
}
\arguments{
\item{block}{An integer block number (as a hexidecimal string) or one of "earliest", "latest" or "pending".}
}
\description{
Returns the number of transactions in a specified block.
}
