% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpc-eth.R
\name{eth_getTokenBalance}
\alias{eth_getTokenBalance}
\title{Returns the tokens balance of the account at specified address.}
\usage{
eth_getTokenBalance(token_address, address)
}
\arguments{
\item{token_address}{A 20 byte (hexadecimal) Token address.}

\item{address}{A 20 byte (hexadecimal) Ethereum address.}
}
\value{
Integer.
}
\description{
Returns the tokens balance of the account at specified address.
}
\examples{
\dontrun{
#' eth_getTokenBalance(token_address = "0x2b591e99afe9f32eaa6214f7b7629768c40eeb39",
# address = "0xb96376d80a16af6700dcbaba2a459dd7856f103a")
}
}
