% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth.R
\name{eth_getTransactionByBlockHashAndIndex}
\alias{eth_getTransactionByBlockHashAndIndex}
\title{Returns information about a transaction by block hash and transaction index position.}
\usage{
eth_getTransactionByBlockHashAndIndex(hash, index)
}
\arguments{
\item{hash}{A block hash.}

\item{index}{Index of transaction in block.}
}
\description{
Returns information about a transaction by block hash and transaction index position.
}
