% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth.R
\name{eth_getTransactionByHash}
\alias{eth_getTransactionByHash}
\title{Returns the information about a transaction requested by transaction hash.}
\usage{
eth_getTransactionByHash(hash)
}
\arguments{
\item{hash}{A transaction hash.}
}
\description{
Returns the information about a transaction requested by transaction hash.
}
