% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth.R
\name{eth_getTransactionCount}
\alias{eth_getTransactionCount}
\title{Title}
\usage{
eth_getTransactionCount(address = NULL, block = "latest")
}
\arguments{
\item{address}{A 20 byte (hexadecimal) Ethereum address.}

\item{block}{An integer block number (as a hexidecimal string) or one of "earliest", "latest" or "pending".}
}
\description{
Title
}
