% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth.R
\name{eth_getTransactionReceipt}
\alias{eth_getTransactionReceipt}
\title{Returns the receipt of a transaction by transaction hash.}
\usage{
eth_getTransactionReceipt(hash)
}
\arguments{
\item{hash}{A transaction hash.}
}
\description{
Returns the receipt of a transaction by transaction hash.
}
