% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth.R
\name{eth_getUncleByBlockHashAndIndex}
\alias{eth_getUncleByBlockHashAndIndex}
\title{Returns information about a uncle of a block by hash and uncle index position.}
\usage{
eth_getUncleByBlockHashAndIndex(hash, index)
}
\arguments{
\item{hash}{Hash of a block.}

\item{index}{Index of transaction in block.}
}
\value{
List.
}
\description{
Returns information about a uncle of a block by hash and uncle index position.
}
\examples{
\dontrun{
eth_getUncleByBlockHashAndIndex(
  "0x8575df1eb3df61f3880628ca8e495038ee0b278c0aa48fe41f80b8d0d4e83e79", "0x0"
)
}
}
