% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth.R
\name{eth_getUncleByBlockNumberAndIndex}
\alias{eth_getUncleByBlockNumberAndIndex}
\title{Returns information about a uncle of a block by number and uncle index position.}
\usage{
eth_getUncleByBlockNumberAndIndex(number, index)
}
\arguments{
\item{number}{An integer block number (as a hexidecimal string) or one of "earliest", "latest" or "pending".}

\item{index}{Index of transaction in block.}
}
\value{
List.
}
\description{
Returns information about a uncle of a block by number and uncle index position.
}
\examples{
\dontrun{
eth_getUncleByBlockNumberAndIndex("0x4720fe", "0x0")
}
}
