% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth.R
\name{eth_getUncleCountByBlockHash}
\alias{eth_getUncleCountByBlockHash}
\title{Returns the number of uncles in a block from a block matching the given block hash.}
\usage{
eth_getUncleCountByBlockHash(hash)
}
\arguments{
\item{hash}{Hash of a block.}
}
\value{
Integer.
}
\description{
Returns the number of uncles in a block from a block matching the given block hash.
}
\examples{
\dontrun{
eth_getUncleCountByBlockHash("0x8575df1eb3df61f3880628ca8e495038ee0b278c0aa48fe41f80b8d0d4e83e79")
}
}
