% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpc-eth.R
\name{eth_getUncleCountByBlockNumber}
\alias{eth_getUncleCountByBlockNumber}
\title{Returns the number of uncles in a block from a block matching the given block number.}
\usage{
eth_getUncleCountByBlockNumber(number)
}
\arguments{
\item{number}{An integer block number (as a hexidecimal string) or one of "earliest", "latest" or "pending".}
}
\value{
Integer.
}
\description{
Returns the number of uncles in a block from a block matching the given block number.
}
\examples{
\dontrun{
eth_getUncleCountByBlockNumber("0x4720fe")
}
}
