% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth.R
\name{eth_getWork}
\alias{eth_getWork}
\title{Returns the hash of the current block, the seedHash, and the boundary condition to be met ("target").}
\usage{
eth_getWork()
}
\description{
Returns the hash of the current block, the seedHash, and the boundary condition to be met ("target").
}
