% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpc-eth.R
\name{eth_mining}
\alias{eth_mining}
\title{Returns true if client is actively mining new blocks.}
\usage{
eth_mining()
}
\value{
Boolean.
}
\description{
Returns true if client is actively mining new blocks.
}
\examples{
\dontrun{
eth_mining()
}
}
