% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth.R
\name{eth_protocolVersion}
\alias{eth_protocolVersion}
\title{Returns the current ethereum protocol version.}
\usage{
eth_protocolVersion()
}
\value{
Integer.
}
\description{
Returns the current ethereum protocol version.
}
\examples{
\dontrun{
eth_protocolVersion()
}
}
