% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth.R
\name{eth_syncing}
\alias{eth_syncing}
\title{Returns an object with data about the sync status.}
\usage{
eth_syncing()
}
\value{
Boolean.
}
\description{
Returns an object with data about the sync status.
}
\examples{
\dontrun{
eth_syncing()
}
}
