% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethereum.R
\docType{package}
\name{ethereum}
\alias{ethereum}
\alias{ethereum-package}
\title{ethereum: An R package for working with Ethereum.}
\description{
The ethereum package provides an interface to the Ethereum RPC functions.
}
