% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{get_blocks}
\alias{get_blocks}
\title{Retrieve a series of blocks.}
\usage{
get_blocks(start = NULL, stop = NULL, count = NULL)
}
\arguments{
\item{start}{Initial block number.}

\item{stop}{Final block number.}

\item{count}{Number of blocks.}
}
\value{
Data frame of blocks.
}
\description{
Retrieve a series of blocks.
}
\examples{
\dontrun{
get_blocks("0x4720FF", "0x472108")
get_blocks("0x4720FF", count = 5)
get_blocks("0x49a8ea", count = 5) # Two blocks containing uncles.
}
}
