% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/www.R
\name{get_post_response}
\alias{get_post_response}
\title{Submit POST to Ethereum RPC.}
\usage{
get_post_response(method, params = list())
}
\arguments{
\item{method}{RPC method.}

\item{params}{List of parameters.}
}
\value{
HTTP response.
}
\description{
We could pass a list as the body argument to POST() and specify encode = "json", but building the JSON package
manually gives us greater control and the ability to compare to results from curl on the command line.
}
\details{
When using curl on command line, you might need to give -H "Content-Type: application/json".
}
\keyword{internal}
