% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{get_transactions}
\alias{get_transactions}
\title{Retrieve transactions for a series of blocks.}
\usage{
get_transactions(start = NULL, stop = NULL, count = NULL)
}
\arguments{
\item{start}{Initial block number.}

\item{stop}{Final block number.}

\item{count}{Number of blocks.}
}
\value{
Data frame of transactions.
}
\description{
Retrieve transactions for a series of blocks.
}
\examples{
\dontrun{
get_transactions("0x4720FF", "0x472108")
get_transactions("0x4720FF", count = 5)
}
}
