% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{hex_to_dec}
\alias{hex_to_dec}
\title{Convert hexadecimal to decimal.}
\usage{
hex_to_dec(hex, integer = TRUE)
}
\arguments{
\item{hex}{Hexadecimal number as string.}

\item{integer}{Whether result should be an integer.}
}
\value{
Numeric or integer.
}
\description{
Convert hexadecimal to decimal.
}
\examples{
hex_to_dec("0x5272")
}
