% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ether-application.R
\name{rollingAverage}
\alias{rollingAverage}
\title{Rolling average}
\usage{
rollingAverage(data, lengthOfRoll)
}
\arguments{
\item{data, }{lengthOfRoll}
}
\value{
rollingAverage
}
\description{
Gets the rolling average for the time period and data specified
}
\examples{
rollingAverage()
}
\keyword{average}
\keyword{rolling,}
\keyword{sma,}
