% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpc-web3.R
\name{web3_clientVersion}
\alias{web3_clientVersion}
\title{Returns the current client version.}
\usage{
web3_clientVersion()
}
\value{
String.
}
\description{
Returns the current client version.
}
\examples{
\dontrun{
web3_clientVersion()
}
}
