% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpc-web3.R
\name{web3_sha3}
\alias{web3_sha3}
\title{Returns Keccak-256 (not the standardized SHA3-256) of the given data.}
\usage{
web3_sha3(data)
}
\arguments{
\item{data}{Data encoded as hex.}
}
\value{
String.
}
\description{
Returns Keccak-256 (not the standardized SHA3-256) of the given data.
}
\examples{
\dontrun{
web3_sha3("0x68656c6c6f20776f726c64")
}
}
