% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth.R
\name{eth_getBalance}
\alias{eth_getBalance}
\title{Returns the balance (in Wei) of the account at specified address.}
\usage{
eth_getBalance(address = NULL, block = "latest")
}
\arguments{
\item{address}{A 20 byte (hexadecimal) Ethereum address.}

\item{block}{An integer block number (as a hexidecimal string) or one of "earliest", "latest" or "pending".}
}
\value{
Integer.
}
\description{
Returns the balance (in Wei) of the account at specified address.
}
\examples{
\dontrun{
eth_getBalance()
eth_getBalance("0xD34DA389374CAAD1A048FBDC4569AAE33fD5a375")
}
}
