% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth.R
\name{eth_getStorageAt}
\alias{eth_getStorageAt}
\title{Returns the value from a storage position at a given address.}
\usage{
eth_getStorageAt(address, position, block = "latest")
}
\arguments{
\item{address}{A 20 byte (hexadecimal) Ethereum address.}

\item{position}{Position in the storage.}

\item{block}{An integer block number (as a hexidecimal string) or one of "earliest", "latest" or "pending".}
}
\value{
Hex string.
}
\description{
Returns the value from a storage position at a given address.
}
\examples{
\dontrun{
eth_getStorageAt("0x3589d05a1ec4Af9f65b0E5554e645707775Ee43C", "0x0")
}
}
