% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpc-eth.R
\name{eth_getTransactionByBlockHashAndIndex}
\alias{eth_getTransactionByBlockHashAndIndex}
\title{Returns information about a transaction by block hash and transaction index position.}
\usage{
eth_getTransactionByBlockHashAndIndex(hash, index)
}
\arguments{
\item{hash}{A block hash.}

\item{index}{Index of transaction in block.}
}
\value{
List.
}
\description{
Returns information about a transaction by block hash and transaction index position.
}
\examples{
\dontrun{
eth_getTransactionByBlockHashAndIndex(
  "0xb6d656ead4c3d4b1aa24d6b4d3d4cde8c090794e597258993512d650f088fcba", "0x0"
)
eth_getTransactionByBlockHashAndIndex(
  "0xb6d656ead4c3d4b1aa24d6b4d3d4cde8c090794e597258993512d650f088fcba", "0x5"
)
}
}
