% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth.R
\name{eth_getTransactionByBlockNumberAndIndex}
\alias{eth_getTransactionByBlockNumberAndIndex}
\title{Returns information about a transaction by block number and transaction index position.}
\usage{
eth_getTransactionByBlockNumberAndIndex(number, index)
}
\arguments{
\item{number}{An integer block number (as a hexidecimal string) or one of "earliest", "latest" or "pending".}

\item{index}{Index of transaction in block.}
}
\value{
List.
}
\description{
Returns information about a transaction by block number and transaction index position.
}
\examples{
\dontrun{
eth_getTransactionByBlockNumberAndIndex("0x4720FF", "0x0")
eth_getTransactionByBlockNumberAndIndex("0x4720FF", "0x5")
}
}
