% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth.R
\name{eth_getTransactionByHash}
\alias{eth_getTransactionByHash}
\title{Returns the information about a transaction.}
\usage{
eth_getTransactionByHash(hash)
}
\arguments{
\item{hash}{A transaction hash.}
}
\value{
List.
}
\description{
Returns the information about a transaction.
}
\examples{
\dontrun{
eth_getTransactionByHash("0x194c67ef1a36990e1aefc7a7ed7855fc2b65c56b588c2ef69e58be2c492a57db")
}
}
