% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth.R
\name{eth_getTransactionByHash}
\alias{eth_getTransactionByHash}
\title{Returns the information about a transaction.}
\usage{
eth_getTransactionByHash(hash)
}
\arguments{
\item{hash}{A transaction hash.}
}
\value{
List.
}
\description{
Returns the information about a transaction.
}
\examples{
\dontrun{
eth_getTransactionByHash("0x4b0f317c326299a4d063ccccff2daf7452c35dc7ec5b4279f001d6e2f745cc6f")
eth_getTransactionByHash("0x8fd9e04958cc4fb602b9d9fa5a9b6da512779ccd22477ccc5ce73721296cf151")
}
}
