% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth.R
\name{eth_getTransactionCount}
\alias{eth_getTransactionCount}
\title{Returns the number of transactions sent from an address.}
\usage{
eth_getTransactionCount(address = NULL, block = "latest")
}
\arguments{
\item{address}{A 20 byte (hexadecimal) Ethereum address.}

\item{block}{An integer block number (as a hexidecimal string) or one of "earliest", "latest" or "pending".}
}
\value{
Integer.
}
\description{
Returns the number of transactions sent from an address.
}
\examples{
eth_getTransactionCount("0xD34DA389374CAAD1A048FBDC4569AAE33fD5a375")
}
