% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth.R
\name{eth_getTransactionReceipt}
\alias{eth_getTransactionReceipt}
\title{Returns the receipt of a transaction by transaction hash.}
\usage{
eth_getTransactionReceipt(hash)
}
\arguments{
\item{hash}{A transaction hash.}
}
\value{
List.
}
\description{
Returns the receipt of a transaction by transaction hash.
}
\examples{
eth_getTransactionReceipt("0x194c67ef1a36990e1aefc7a7ed7855fc2b65c56b588c2ef69e58be2c492a57db")
}
