% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpc-eth.R
\name{eth_getTransactionReceipt}
\alias{eth_getTransactionReceipt}
\title{Returns the receipt of a transaction by transaction hash.}
\usage{
eth_getTransactionReceipt(hash)
}
\arguments{
\item{hash}{A transaction hash.}
}
\value{
List.
}
\description{
Returns the receipt of a transaction by transaction hash.
}
\examples{
\dontrun{
eth_getTransactionReceipt("0x8fd9e04958cc4fb602b9d9fa5a9b6da512779ccd22477ccc5ce73721296cf151")
}
}
