% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth.R
\name{eth_getUncleCountByBlockHash}
\alias{eth_getUncleCountByBlockHash}
\title{Returns the number of uncles in a block from a block matching the given block hash.}
\usage{
eth_getUncleCountByBlockHash(hash)
}
\arguments{
\item{hash}{Hash of a block.}
}
\description{
Returns the number of uncles in a block from a block matching the given block hash.
}
