% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpc-eth.R
\name{eth_hashrate}
\alias{eth_hashrate}
\title{Returns the number of hashes per second that the node is mining with.}
\usage{
eth_hashrate()
}
\value{
Integer.
}
\description{
Returns the number of hashes per second that the node is mining with.
}
\examples{
\dontrun{
eth_hashrate()
}
}
