% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ether-application.R
\name{priceDistribution}
\alias{priceDistribution}
\title{Distribution}
\usage{
priceDistribution(data, distribution)
}
\arguments{
\item{data, }{distribution}
}
\description{
Gets the prcing distribution of a dataset
}
\examples{
rollingAverage()
}
